examples/supervised/handwriting/requirements.txt
imageio
pillow

# We will use tensorflow 1.14, because for releases 1.14 and older, CPU and GPU packages are separate
# tensorflow==1.14 —Release for CPU-only
# tensorflow-gpu==1.14 —Release with GPU support (Ubuntu and Windows)

# Also use an older numpy version to not get the future warnings from deprecated functions in newest numpy from
# calls from the older tensorflow.
numpy==1.16.6

# TW: Setuptools 44.0 is available, but does not work with Tensorflow 1.14
setuptools==41.0


tensorflow==1.14
